refactor(validation): update import paths and improve error handling in ZodHttpValidator
