--- backend/src/main/modules/domain/usecases/create-user.ts
@@ -1,11 +1,0 @@
-import type { User } from "../entities/user";
-
-export interface CreateUserDTO {
-  email: string;
-  name: string;
-  password: string;
-}
-
-export interface CreateUser {
-  execute(user: CreateUserDTO): Promise<User>;
-}
