+++ backend/src/main/modules/presentation/validation/zod-http-validator.ts
@@ -1,10 +1,10 @@
-import { AppError } from "@presentation/errors/app-error";
-import type { CreateUserDTO } from "@domain/usecases/create-user";
-import type { HttpRequest } from "@presentation/dtos/http-request.dto";
-import type { createUserSchema } from "./create-user.schema";
-import type { HttpValidator } from "@presentation/contracts/http-validator.contract";
 import type { z } from "zod";
+import type { CreateUserDTO } from "@domain/usecases/create-user.contract";
+import type { HttpRequest } from "@presentation/dtos/http-request.dto";
+import type { createUserSchema } from "@presentation/validation/create-user.schema";
+import type { HttpValidator } from "@presentation/contracts/http-validator.contract";
 import type { ERROR_CODE } from "@presentation/errors/error-codes";
+import { AppError } from "@presentation/errors/app-error";
 
 type Schema = z.infer<typeof createUserSchema>;
 
@@ -13,14 +13,17 @@ export class ZodHttpValidator implements HttpValidator {
 
   execute(request: HttpRequest): CreateUserDTO {
     const parsed = this.schema.safeParse(request);
-    let detail: string | undefined;
 
     if (!parsed.success) {
-      const error = parsed.error.issues[0].message as ERROR_CODE;
+      const issue = parsed.error.issues[0];
+      const error = issue.message as ERROR_CODE;
+      const issueField = issue.path.at(-1)?.toString();
+
       if (error === "MISSING_PARAM" || error === "INVALID_PARAM") {
-        detail = parsed.error.issues[0].path.at(-1)?.toString();
+        throw new AppError(error, issueField);
       }
-      throw new AppError(error, detail);
+
+      throw new AppError(error);
     }
 
     const userDTO = (parsed.data as Schema).body;
